fix(VarHolder): MSVC warning #4886 (#4903)
* fix(VarHolder): MSVC warning #4886

* fix(VarHolder): do not convert unsigned #4886

* chore(SocketImpl): eliminate MSVC warning

* fix(VarHolder): do not convert unsigned (extract unpreserveSign()) #4886

* chore(testsuite): eliminate MSVC warnings

* chore(VarHolder): simplify unpreserveSign using constexpr.

---------

+++ Foundation/include/Poco/String.h
@@ -400,7 +400,7 @@ int icompare(
 	typename S::size_type pos,
 	const typename S::value_type* ptr)
 {
-	int n = pos < str.size() ? str.size() - pos : 0;
+	int n = static_cast<int>(pos < str.size() ? str.size() - pos : 0);
 	return icompare(str, pos, n, ptr);
 }
